store/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
python-multipart
pyjwt
cryptography
spacy
gensim
transformers
torch
requests
python-dotenv
